create useButton hook
