+++ src/app/components/custom-input/hooks/use-input.hook.js
@@ -1,0 +1,15 @@
+import { useState } from "react";
+
+
+export const useInput = (startValue) => {
+    const [value, setValue] = useState(startValue);
+    const handleChange = (event) => {
+            setValue(event.target.value)
+        }
+        console.log('value', value)
+        return {
+            value,
+            setValue,
+            handleChange,
+        }
+    }
